Move file source models to module
And add generic types and utils for configuration template resolution
