+++ lib/galaxy/files/plugins.py
@@ -1,15 +1,9 @@
 from typing import (
     cast,
-    Optional,
     TYPE_CHECKING,
 )
 
-from pydantic import BaseModel
-
-from galaxy.util.config_parsers import (
-    IpAllowedListEntryT,
-    parse_allowlist_ips,
-)
+from galaxy.files.models import FileSourcePluginsConfig
 from galaxy.util.plugin_config import (
     load_plugins,
     PluginConfigSource,
@@ -18,63 +12,6 @@ from galaxy.util.plugin_config import (
 
 if TYPE_CHECKING:
     from galaxy.files.sources import BaseFilesSource
-
-
-class FileSourcePluginsConfig(BaseModel):
-    symlink_allowlist: list[str] = []
-    fetch_url_allowlist: list[IpAllowedListEntryT] = []
-    library_import_dir: Optional[str] = None
-    user_library_import_dir: Optional[str] = None
-    ftp_upload_dir: Optional[str] = None
-    ftp_upload_purge: bool = True
-    tmp_dir: Optional[str] = None
-    webdav_use_temp_files: Optional[bool] = None
-    listings_expiry_time: Optional[int] = None
-
-    @staticmethod
-    def from_app_config(config):
-        # Formalize what we read in from config to create a more clear interface
-        # for this component.
-        kwds = {}
-        kwds["symlink_allowlist"] = config.user_library_import_symlink_allowlist
-        kwds["fetch_url_allowlist"] = [str(ip) for ip in config.fetch_url_allowlist_ips]
-        kwds["library_import_dir"] = config.library_import_dir
-        kwds["user_library_import_dir"] = config.user_library_import_dir
-        kwds["ftp_upload_dir"] = config.ftp_upload_dir
-        kwds["ftp_upload_purge"] = config.ftp_upload_purge
-        kwds["tmp_dir"] = config.file_source_temp_dir
-        kwds["webdav_use_temp_files"] = config.file_source_webdav_use_temp_files
-        kwds["listings_expiry_time"] = config.file_source_listings_expiry_time
-
-        return FileSourcePluginsConfig(**kwds)
-
-    def to_dict(self):
-        return {
-            "symlink_allowlist": self.symlink_allowlist,
-            "fetch_url_allowlist": self.fetch_url_allowlist,
-            "library_import_dir": self.library_import_dir,
-            "user_library_import_dir": self.user_library_import_dir,
-            "ftp_upload_dir": self.ftp_upload_dir,
-            "ftp_upload_purge": self.ftp_upload_purge,
-            "tmp_dir": self.tmp_dir,
-            "webdav_use_temp_files": self.webdav_use_temp_files,
-            "listings_expiry_time": self.listings_expiry_time,
-        }
-
-    @staticmethod
-    def from_dict(as_dict):
-        return FileSourcePluginsConfig(
-            symlink_allowlist=as_dict["symlink_allowlist"],
-            fetch_url_allowlist=parse_allowlist_ips(as_dict["fetch_url_allowlist"]),
-            library_import_dir=as_dict["library_import_dir"],
-            user_library_import_dir=as_dict["user_library_import_dir"],
-            ftp_upload_dir=as_dict["ftp_upload_dir"],
-            ftp_upload_purge=as_dict["ftp_upload_purge"],
-            # Always provided for new jobs, remove in 25.0
-            tmp_dir=as_dict.get("tmp_dir"),
-            webdav_use_temp_files=as_dict.get("webdav_use_temp_files"),
-            listings_expiry_time=as_dict.get("listings_expiry_time"),
-        )
 
 
 class FileSourcePluginLoader:
