+++ lib/galaxy/files/sources/_defaults.py
@@ -1,0 +1,2 @@
+DEFAULT_WRITABLE = False
+DEFAULT_SCHEME = "gxfiles"
