+++ lib/galaxy/files/sources/util.py
@@ -6,10 +6,8 @@ from galaxy.files import (
     ConfiguredFileSources,
     FileSourcesUserContext,
 )
-from galaxy.files.sources import (
-    FilesSourceOptions,
-    PartialFilesSourceProperties,
-)
+from galaxy.files.models import PartialFilesSourceProperties
+from galaxy.files.sources import FilesSourceOptions
 from galaxy.files.uris import stream_url_to_file
 from galaxy.util import (
     DEFAULT_SOCKET_TIMEOUT,
